Improvements after first actual test (only TTbarJets_SemiLept)

+++ Events_MC/AddSFToWeights.py
@@ -28,10 +28,10 @@ if len(sys.argv) > 1:
         if subdir.startswith('Reco_'):
             print "   - ", subdir
 
-sampleNames = {'TTbarJets_FullHadr', 'TTbarJets_FullLept', 'TTbarJets_SemiLept', 'SingleTop_tWChannel_t',
+sampleNames = ['TTbarJets_FullHadr', 'TTbarJets_FullLept', 'TTbarJets_SemiLept', 'SingleTop_tWChannel_t',
                'SingleTop_tWChannel_tbar', 'SingleTop_tChannel_t', 'SingleTop_tChannel_tbar', 'SingleTop_sChannel_t',
                'SingleTop_sChannel_tbar', 'WJets_1jets', 'WJets_2jets', 'WJets_3jets', 'WJets_4jets', 'ZJets_1jets',
-               'ZJets_2jets', 'ZJets_3jets', 'ZJets_4jets'}
+               'ZJets_2jets', 'ZJets_3jets', 'ZJets_4jets']
 whichSample = "-1"
 if len(sys.argv) > 2:
     whichSample = sys.argv[2]
@@ -43,13 +43,18 @@ if len(sys.argv) > 3:
     CWUChoice = sys.argv[3]
     print "Specified whether Correct/Wrong/Unmatched TTbarJets_SemiLept events should be considered!!"
 
-Luminosity, NormFactor, NrEvts, consSamples = [], [], [], []
-EvtNrMatchingArray = [[]]
-for iSample in range(len(sampleNames)):
-    if (str(whichSample) == "-1") or (str(whichSample) != "-1" and str(whichSample) in sampleNames[iSample]):
+if str(whichSample) != "-1" and str(whichSample) != "TTbarJets_SemiLept":
+    CWUChoice = "0"
 
-        EvtNrMatching = open('MWEventNrMatching_'+sampleNames[iSample]+'.txt', 'r')
+Luminosity, NormFactor, NrEvts, consSamples = [], [], [], []
+EvtNrMatchingArray = [[],[],[],[],[],[],[],[],[],[],[],[],[],[],[],[],[],[],[],[]]
+for iSample in range(len(sampleNames)):
+    print "Looking at : ", sampleNames[iSample]
+    if (str(whichSample) == "-1") or (str(whichSample) != "-1" and str(whichSample) in sampleNames[iSample]):
+        print " --> Will continue with this one ! \n"
+
         if str(sampleNames[iSample]) != "TTbarJets_SemiLept":
+            EvtNrMatching = open('MWEventNrMatching_'+sampleNames[iSample]+'.txt', 'r')
             consSamples.append(str(sampleNames[iSample]))   # Store the name of the considered sample for weights file!
             sampleSize = len(EvtNrMatchingArray)
             for evtNr in EvtNrMatching:
@@ -67,45 +72,57 @@ for iSample in range(len(sampleNames)):
                     EvtNrMatchingArray[len(consSamples)-1].append(str(evtNrWord[6]))
             # NrEvts.append(len(EvtNrMatching)-sampleSize)
             print "Stored a total of ", len(EvtNrMatchingArray[len(consSamples)-1]), " events for ", sampleNames[iSample], "!"
+            EvtNrMatching.close()
         else:
             # Special case for TTbarJets since this has been splitted in correct/wrong/unmatched!
+            print " ... is semiLept thus separate index for correct/wrong/unmatched ..."
 
             for iCWU in range(len(CWUOptions)):
+                print " Looking at CWU index of : ", CWUOptions[iCWU]
                 if str(CWUChoice) == "-1" or (str(CWUChoice) != "-1" and str(CWUChoice) in CWUOptions):
-                    print "Now looking at ", NameOptions[iCWU], "events !"
+                    print " --> Now looking at ", CWUOptions[iCWU], "events ! \n"
 
+                    EvtNrMatching = open('MWEventNrMatching_'+sampleNames[iSample]+'.txt', 'r')
                     consSamples.append(str(sampleNames[iSample])+'_'+CWUOptions[iCWU])
+                    print "Array with the considered samples has ", len(consSamples), " entries : ", consSamples
                     for evtNrTT in EvtNrMatching:
                         evtNrWordTT = evtNrTT.split()
 
                         #First get the lumi and the normFactor!
                         if re.search(r" * Lumi = ", evtNrTT):
                             Luminosity.append(evtNrWordTT[3])
-                            print "Luminosity is : ", evtNrWordTT[3]
+                            print " Luminosity is : ", evtNrWordTT[3]
+                            print " Luminosity array has ", len(Luminosity), " entries : ", Luminosity
                         if re.search(r" * NormFactor = ", evtNrTT):
                             NormFactor.append(evtNrWordTT[3])
                             print "Norm factor is : ", evtNrWordTT[3]
+                            print " Norm factoar array has ", len(NormFactor), " entries : ", NormFactor
 
                         if len(evtNrWordTT) == 7 and str(evtNrWordTT[4]) == CWUOptions[iCWU]:
                             EvtNrMatchingArray[len(consSamples)-1].append(str(evtNrWordTT[6]))
-        EvtNrMatching.close()
-    else:
-        print "Issue with sample name ... --> ", whichSample
+                    print "Filled EvtNrMatching array with ", len(EvtNrMatchingArray[len(consSamples)-1]), " entries ... :", EvtNrMatchingArray[len(consSamples)-1]
+                    EvtNrMatching.close()
 
+print " "
 # -- Now go into the different directories and change the weights.out file -- #
-for mcDir in os.listdir():
-    if (str(whichDir) == "-1" and not 'Combined' in mcDir and not '.' in mcDir) or (str(whichDir) != "-1" and str(whichDir) in dirNames):
-        combinedMCWeights = open('weights_CheckedEvts_CombinedMC_SFAdded.txt', 'w')  # One output file for all MC!
+combinedMCWeights = open('weights_CheckedEvts_CombinedMC_SFAdded.out', 'w')  # One output file for all MC!
+for mcDir in os.listdir("."):
+    if (str(whichDir) == "-1" and not 'Combined' in mcDir and not '.' in mcDir) or (str(whichDir) != "-1" and str(whichDir) in mcDir and not '.' in mcDir):
+        combinedSampleWeights = open(mcDir+'/weights_CheckedEvts_Combined'+mcDir+'_SFAdded.out', 'w')
+        combinedSemiMuWeights= " "  # = open(mcDir+'/weights_CheckedEvts_CombinedSemiLeptEvts_SFAdded.out', 'w')
 
         print "Looking at directory : ", mcDir
+        if str(mcDir) == "TTbarJets":
+            combinedSemiMuWeights = open(mcDir+'/weights_CheckedEvts_CombinedSemiLeptEvts_SFAdded.out', 'w')
+
         for MCsubdir in os.listdir(mcDir):
-            if (str(whichSample) == "-1" and MCsubdir.startswith('Reco_')) or (str(whichSample) != "-1" and str(whichSample) in sampleNames):
-                combinedSampleWeights = open(mcDir+'/weights_CheckedEvts_Combined'+mcDir+'_SFAdded.txt')
+            if (str(whichSample) == "-1" and MCsubdir.startswith('Reco_')) or (str(whichSample) != "-1" and str(whichSample) in MCsubdir):
+                print " --> Continuing with directory ", MCsubdir
 
                 if not 'TTbarJets_SemiLept' in MCsubdir:
 
-                    origWeights = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts.txt')
-                    newWeights = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts_SFAdded.txt')
+                    origWeights = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts.out', 'r')
+                    newWeights = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts_SFAdded.out', 'w')
 
                     # Get the correct index for the filled arrays!
                     arrayIndex = 999
@@ -126,13 +143,13 @@ for mcDir in os.listdir():
                     newWeights.close()
                 else:
                     # Special case for semiLept events because they are splitted in Correct/Wrong/Unmatched!
-                    combinedSemiMuWeights = open(mcDir+'/weights_CheckedEvts_CombinedSemiMuEvts_SFAdded.txt')
 
-                    for iCase in range(len(CWUChoice)):
+                    for iCase in range(len(CWUOptions)):
+                        print "  * Looping over the different CWU options ", CWUOptions[iCase]
                         if str(CWUChoice) == "-1" or (str(CWUChoice) != "-1" and str(CWUChoice) in CWUOptions):
-                            if CWUChoice[iCase] in MCsubdir:
-                                origWeightsTT = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts.txt')
-                                newWeightsTT = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts_SFAdded.txt')
+                            if CWUOptions[iCase] in MCsubdir:
+                                origWeightsTT = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts.out', 'r')
+                                newWeightsTT = open(mcDir+'/'+MCsubdir+'/weights_CheckedEvts_SFAdded.out', 'w')
 
                                 # Get the correct index for the filled arrays!
                                 arrayIndex = 999
@@ -140,31 +157,34 @@ for mcDir in os.listdir():
                                     if MCsubdir.startswith('Reco_'+str(j)):
                                         arrayIndex = int(i)
 
-                                for line in origWeights:
+                                for line in origWeightsTT:
                                     word = line.split()
 
                                     if str(word[0]) != "#" and int(word[0]) <= 117658:
-                                        newWeights.write(word[0]+' '+word[1]+' '+word[2]+' '+word[3]+' '+word[4]+' '+EvtNrMatchingArray[arrayIndex][int(word[0])-1]+' '+Luminosity[arrayIndex]+' '+NormFactor[arrayIndex]+' '+consSamples[arrayIndex]+'\n')
+                                        newWeightsTT.write(word[0]+' '+word[1]+' '+word[2]+' '+word[3]+' '+word[4]+' '+EvtNrMatchingArray[arrayIndex][int(word[0])-1]+' '+Luminosity[arrayIndex]+' '+NormFactor[arrayIndex]+' '+consSamples[arrayIndex]+'\n')
                                         if str(whichSample) == "-1":
                                             combinedSampleWeights.write(word[0]+' '+word[1]+' '+word[2]+' '+word[3]+' '+word[4]+' '+EvtNrMatchingArray[arrayIndex][int(word[0])-1]+' '+Luminosity[arrayIndex]+' '+NormFactor[arrayIndex]+' '+consSamples[arrayIndex]+'\n')
                                         if str(CWUChoice) == "-1":
                                             combinedSemiMuWeights.write(word[0]+' '+word[1]+' '+word[2]+' '+word[3]+' '+word[4]+' '+EvtNrMatchingArray[arrayIndex][int(word[0])-1]+' '+Luminosity[arrayIndex]+' '+NormFactor[arrayIndex]+' '+consSamples[arrayIndex]+'\n')
                                     elif str(word[0]) == "#":
-                                        newWeights.write(line)
-                            origWeightsTT.close()
-                            newWeightsTT.close()
+                                        newWeightsTT.write(line)
+                                origWeightsTT.close()
+                                newWeightsTT.close()
 
-                    # Remove this combined SemiLept weights output file in case only 1 CWU type has been considered
-                    combinedSemiMuWeights.close()
-                    if str(CWUChoice) != "-1":
-                        os.system('rm '+combinedSemiMuWeights.name)
+        # Remove this combined SemiLept weights output file in case only 1 CWU type has been considered
+        if str(mcDir) == "TTbarJets":
+            combinedSemiMuWeights.close()
+            if str(CWUChoice) != "-1":
+                os.system('rm '+combinedSemiMuWeights.name)
 
-                # Remove this combined MC type weights output file in case only 1 MC sample has been considered
-                combinedSampleWeights.close()
-                if str(whichSample) != "-1":
-                    os.system('rm '+combinedSampleWeights.name)
+        # Remove this combined MC type weights output file in case only 1 MC sample has been considered
+        combinedSampleWeights.close()
+        if str(whichSample) != "-1":
+            os.system('rm '+combinedSampleWeights.name)
+    else:
+        print "    **** Nothing done with directory ", mcDir
 
-        # Remove this combined weights output file in case only 1 MC directory has been considered!
-        combinedMCWeights.close()
-        if str(whichDir) != "-1":
-            os.system('rm '+combinedMCWeights.name)
+# Remove this combined weights output file in case only 1 MC directory has been considered!
+combinedMCWeights.close()
+if str(whichDir) != "-1":
+    os.system('rm '+combinedMCWeights.name)
